backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.29.0
sqlalchemy==2.0.30
psycopg2-binary==2.9.9
pydantic==2.7.1
pydantic-settings==2.3.0
python-jose[cryptography]==3.3.0
httpx==0.27.0
icalendar==5.0.12
python-multipart==0.0.9
google-auth==2.29.0
email-validator==2.1.1
alembic==1.13.1
